fix(pages): buttons actions in speaker details
