+++ src/shared/locales/en/general.json
@@ -75,5 +75,6 @@
   "tr_featureComingSoonDesc": "This feature is not available yet. We are currently working on it. Stay tuned!",
   "tr_openRealApp": "Open real App",
   "tr_confirm": "Confirm",
-  "tr_hide": "Hide"
+  "tr_hide": "Hide",
+  "tr_close": "Close"
 }
