+++ src/v3/features/persons/visiting_speakers/speaker_details/contact_info/index.tsx
@@ -1,17 +1,18 @@
 import { Box } from '@mui/material';
+import { IconCall, IconMail } from '@components/icons';
 import { useAppTranslation } from '@hooks/index';
 import { SpeakerContactInfoType } from './index.types';
 import useContactInfo from './useContactInfo';
+import Button from '@components/button';
 import Typography from '@components/typography';
-import { IconCall, IconMail } from '@components/icons';
 
-const SpeakerContactInfo = ({ speaker }: SpeakerContactInfoType) => {
+const SpeakerContactInfo = ({ speaker, onClose }: SpeakerContactInfoType) => {
   const { t } = useAppTranslation();
 
   const { person } = useContactInfo(speaker);
 
   return (
-    <Box sx={{ display: 'flex', flexDirection: 'column', gap: '16px', marginBottom: '24px' }}>
+    <Box sx={{ display: 'flex', flexDirection: 'column', gap: '16px' }}>
       <Box sx={{ display: 'flex', alignItems: 'center', gap: '16px' }}>
         <Box sx={{ display: 'flex', alignItems: 'center', gap: '8px', width: '126px', minWidth: '126px' }}>
           <IconCall color="var(--black)" />
@@ -30,6 +31,10 @@ const SpeakerContactInfo = ({ speaker }: SpeakerContactInfoType) => {
           {person.email.value}
         </Typography>
       </Box>
+
+      <Button variant="main" onClick={onClose} sx={{ width: '100%' }}>
+        {t('tr_close')}
+      </Button>
     </Box>
   );
 };
