+++ src/v3/features/persons/visiting_speakers/speaker_details/contact_info/index.types.ts
@@ -2,4 +2,5 @@ import { VisitingSpeakerType } from '@definition/visiting_speakers';
 
 export type SpeakerContactInfoType = {
   speaker: VisitingSpeakerType;
+  onClose: VoidFunction;
 };
