+++ src/v3/features/persons/visiting_speakers/speaker_details/talks_songs/index.types.ts
@@ -2,5 +2,5 @@ import { VisitingSpeakerType } from '@definition/visiting_speakers';
 
 export type SpeakerTalksSongsType = {
   speaker: VisitingSpeakerType;
-  edit: boolean;
+  onClose: VoidFunction;
 };
