+++ src/v3/features/persons/visiting_speakers/speaker_details/useSpeakerDetails.tsx
@@ -11,17 +11,13 @@ const useSpeakerDetails = (speaker: VisitingSpeakerType) => {
   const persons = useRecoilValue(personsActiveState);
   const congName = useRecoilValue(congNameState);
 
-  const [isEdit, setIsEdit] = useState(false);
-
   const person = persons.find((record) => record.person_uid === speaker.person_uid);
 
   const personName = buildPersonFullname(speaker.person_lastname.value, speaker.person_firstname.value, fullnameOption);
 
   const isElder = personIsElder(person);
 
-  const handleToggleEdit = () => setIsEdit((prev) => !prev);
-
-  return { personName, isElder, congName, handleToggleEdit, isEdit };
+  return { personName, isElder, congName };
 };
 
 export default useSpeakerDetails;
